requirements.txt
flask
dateparser
pypdf
python-docx
Pillow
pytesseract
